api/requirements.txt
crewai
openai
fastapi
uvicorn
python-dotenv